# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is the API for the Green Ecolution Management System.",
        "title": "Green Space Management API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "Green Ecolution",
            "url": "https://green-ecolution.de",
            "email": "info@green-ecolution.de"
        },
        "license": {
            "name": "AGPL",
            "url": "https://raw.githubusercontent.com/green-ecolution/green-ecolution/develop/LICENSE"
        },
        "version": "develop"
    },
    "paths": {
...
        "/v1/region": {
            "get": {
                "security": [
                    {
                        "Keycloak": []
                    }
                ],
                "description": "Get all regions",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Region"
                ],
                "summary": "Get all regions",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Page",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Limit",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/RegionList"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/HTTPError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/HTTPError"
                        }
                    }
                }
            }
        },
        "/v1/region/{id}": {
            "get": {
                "security": [
                    {
                        "Keycloak": []
                    }
                ],
                "description": "Get a region by ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Region"
                ],
                "summary": "Get a region by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Region ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/Region"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/HTTPError"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/HTTPError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/HTTPError"
                        }
                    }
                }
            }
        },
        "/v1/sensor": {
            "get": {
                "security": [
                    {
                        "Keycloak": []
                    }
                ],
                "description": "Get all sensors",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Sensor"
                ],
                "summary": "Get all sensors",
                "operationId": "get-all-sensors",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Page",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Limit",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Provider",
                        "name": "provider",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/SensorList"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/HTTPError"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/HTTPError"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/HTTPError"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/HTTPError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/HTTPError"
                        }
                    }
                }
            }
        },
        "/v1/sensor/data/{sensor_id}": {
            "get": {
                "security": [
                    {
                        "Keycloak": []
                    }
                ],
                "description": "Get all sensor data by id",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Sensor"
                ],
                "summary": "Get all sensor data by id",
                "operationId": "get-all-sensor-data-by-id",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Sensor ID",
                        "name": "sensor_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/SensorDataList"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/HTTPError"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/HTTPError"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/HTTPError"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/HTTPError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/HTTPError"
                        }
                    }
                }
            }
        },
        "/v1/sensor/{sensor_id}": {
            "get": {
                "security": [
                    {
                        "Keycloak": []
                    }
                ],
                "description": "Get sensor by ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Sensor"
                ],
                "summary": "Get sensor by ID",
                "operationId": "get-sensor-by-id",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Sensor ID",
                        "name": "sensor_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/Sensor"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/HTTPError"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/HTTPError"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/HTTPError"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/HTTPError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/HTTPError"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "Keycloak": []
                    }
                ],
                "description": "Delete sensor",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Sensor"
                ],
                "summary": "Delete sensor",
                "operationId": "delete-sensor",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Sensor ID",
                        "name": "sensor_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/HTTPError"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/HTTPError"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/HTTPError"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/HTTPError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/HTTPError"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "HTTPError": {
            "type": "object",
            "required": [
                "code",
                "error",
                "method",
                "path"
            ],
            "properties": {
                "code": {
                    "type": "integer"
                },
                "error": {
                    "type": "string"
                },
                "method": {
                    "type": "string"
                },
                "path": {
                    "type": "string"
                }
            }
        },
...
        "Pagination": {
            "type": "object",
            "required": [
                "current_page",
                "next_page",
                "prev_page",
                "total_pages",
                "total_records"
            ],
            "properties": {
                "current_page": {
                    "type": "integer"
                },
                "next_page": {
                    "type": "integer"
                },
                "prev_page": {
                    "type": "integer"
                },
                "total_pages": {
                    "type": "integer"
                },
                "total_records": {
                    "type": "integer"
                }
            }
        },
...
        "Region": {
            "type": "object",
            "required": [
                "id",
                "name"
            ],
            "properties": {
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                }
            }
        },
...
        "RegionList": {
            "type": "object",
            "required": [
                "data"
            ],
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/Region"
                    }
                },
                "pagination": {
                    "$ref": "#/definitions/Pagination"
                }
            }
        },
...
        "Sensor": {
            "type": "object",
            "required": [
                "created_at",
                "id",
                "latest_data",
                "latitude",
                "longitude",
                "status",
                "updated_at"
            ],
            "properties": {
                "additional_information": {
                    "type": "object",
                    "additionalProperties": true
                },
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "latest_data": {
                    "$ref": "#/definitions/SensorData"
                },
                "latitude": {
                    "type": "number"
                },
                "longitude": {
                    "type": "number"
                },
                "provider": {
                    "type": "string"
                },
                "status": {
                    "$ref": "#/definitions/SensorStatus"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "SensorData": {
            "type": "object",
            "required": [
                "battery",
                "created_at",
                "humidity",
                "temperature",
                "updated_at",
                "watermarks"
            ],
            "properties": {
                "battery": {
                    "type": "number"
                },
                "created_at": {
                    "type": "string"
                },
                "humidity": {
                    "type": "number"
                },
                "temperature": {
                    "type": "number"
                },
                "updated_at": {
                    "type": "string"
                },
                "watermarks": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/WatermarkResponse"
                    }
                }
            }
        },
        "SensorDataList": {
            "type": "object",
            "required": [
                "data"
            ],
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/SensorData"
                    }
                }
            }
        },
        "SensorList": {
            "type": "object",
            "required": [
                "data"
            ],
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/Sensor"
                    }
                },
                "pagination": {
                    "$ref": "#/definitions/Pagination"
                }
            }
        },
        "SensorStatus": {
            "type": "string",
            "enum": [
                "online",
                "offline",
                "unknown"
            ],
            "x-enum-varnames": [
                "SensorStatusOnline",
                "SensorStatusOffline",
                "SensorStatusUnknown"
            ]
        },
...
        "WatermarkResponse": {
            "type": "object",
            "required": [
                "centibar",
                "depth",
                "resistance"
            ],
            "properties": {
                "centibar": {
                    "type": "integer"
                },
                "depth": {
                    "type": "integer"
                },
                "resistance": {
                    "type": "integer"
                }
            }
        }
    },
    "securityDefinitions": {
        "Keycloak": {
            "type": "oauth2",
            "flow": "accessCode",
            "authorizationUrl": "https://auth.green-ecolution.de/realms/green-ecolution-dev/protocol/openid-connect/auth",
            "tokenUrl": "https://auth.green-ecolution.de/realms/green-ecolution-dev/protocol/openid-connect/token"
        }
    }
}